add user detail actvitiy

+++ data/src/main/java/com/example/data/repository/UserRemoteMediator.kt
@@ -10,6 +10,7 @@ import com.example.data.datasource.remote.UserRemoteDataSource
 import com.example.data.entity.QueryEntity
 import com.example.data.entity.UserEntity
 import retrofit2.HttpException
+import timber.log.Timber
 import java.io.IOException
 import javax.inject.Inject
 
@@ -19,7 +20,7 @@ internal class UserRemoteMediator @Inject constructor(
     private val userDataSource: UserLocalDataSource,
     private val remoteDataSource: UserRemoteDataSource,
     private val query: String,
-    private val refresh: Boolean
+    private val isRefresh: Boolean
 ) : RemoteMediator<Int, UserEntity>() {
 
     private var page: Int = queryDataSource.getPage(query) ?: START_PAGE
@@ -33,41 +34,42 @@ internal class UserRemoteMediator @Inject constructor(
                 LoadType.PREPEND ->
                     return MediatorResult.Success(endOfPaginationReached = true)
                 LoadType.APPEND ->
-                    if (page != END_PAGE) ++page else
+                    if (page != LAST_PAGE) page else
                         return MediatorResult.Success(endOfPaginationReached = true)
                 LoadType.REFRESH ->
-                    if (refresh) START_PAGE else
-                        return MediatorResult.Success(endOfPaginationReached = page == END_PAGE)
+                    if (isRefresh) START_PAGE else
+                        return MediatorResult.Success(endOfPaginationReached = page == LAST_PAGE)
             }
 
             val response = remoteDataSource.searchUser(query, queryPage)
 
             response.run {
                 if (isSuccessful) {
-                    val isLastPage = headers()[HEADER_PAGE_KEY]?.contains(HEADER_PAGE_LAST) ?: false
+
+                    val isLastPage = headers()[HEADER_PAGE_KEY]?.contains(HEADER_PAGE_LAST)?.not() ?: false
 
                     body()?.let { response ->
 
-                        val isRefresh = loadType == LoadType.REFRESH && refresh
-                        if (isRefresh) {
+                        if (loadType == LoadType.REFRESH && isRefresh) {
                             userDataSource.deleteByQuery(query)
                         }
-                        userDataSource.insert(response.items.map { user ->
-                            user.apply { this.query = this@UserRemoteMediator.query }
-                        })
+
+                        page = if (isLastPage) LAST_PAGE else queryPage + 1
 
                         queryDataSource.insert(
                             QueryEntity(
-                                query, when {
-                                    isLastPage -> END_PAGE
-                                    isRefresh -> START_PAGE + 1
-                                    else -> page
-                                }
+                                query = query, page = page
                             )
                         )
+
+                        userDataSource.insert(response.items.map { user ->
+                            user.apply { this.query = this@UserRemoteMediator.query }
+                        })
                     }
                     if (isLastPage)
                         return MediatorResult.Success(endOfPaginationReached = true)
+                } else {
+                    return MediatorResult.Error(IOException(response.message()))
                 }
             }
             return MediatorResult.Success(
@@ -82,7 +84,7 @@ internal class UserRemoteMediator @Inject constructor(
 
     companion object {
         const val START_PAGE = 1
-        const val END_PAGE = -1
+        const val LAST_PAGE = -1
 
         const val HEADER_PAGE_KEY = "link"
         const val HEADER_PAGE_LAST = "last"
